scripts/ai-blog-generator/requirements.txt
google-genai>=1.0.0
Pillow>=10.0.0
python-slugify>=8.0.0
